# src/test/resources/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "WEX - Purchase",
    "description" : "This API helps manage all purchases made by WEX users",
    "contact" : {
      "name" : "WEX",
      "url" : "https://wex.io",
      "email" : "contato@wex.io"
    },
    "version" : "1"
  },
  "paths" : {
    "/v1/purchase/{uuid}" : {
      "get" : {
        "summary" : "Retrive a purchase",
        "operationId" : "get",
        "parameters" : [ {
          "name" : "uuid",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "country",
          "in" : "query",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Return the purchase",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PurchaseResponseDto"
                }
              }
            }
          },
          "404" : {
            "description" : "Entity not found",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/v1/purchase" : {
      "post" : {
        "summary" : "Save a new purchase",
        "operationId" : "save",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PurchaseRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Return the saved purchase",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PurchaseResponseDto"
                }
              }
            }
          },
          "422" : {
            "description" : "Unprocessable Entity",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "PurchaseResponseDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "description" : {
            "type" : "string"
          },
          "transactionDate" : {
            "type" : "string",
            "format" : "date"
          },
          "exchangeRate" : {
            "type" : "number"
          },
          "amount" : {
            "type" : "number"
          },
          "convertedMoney" : {
            "type" : "number"
          }
        }
      },
      "ErrorDetails" : {
        "type" : "object",
        "properties" : {
          "target" : {
            "type" : "string"
          },
          "field" : {
            "type" : "string"
          },
          "violationMessage" : {
            "type" : "string"
          }
        }
      },
      "ErrorResponseDto" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "status" : {
            "type" : "integer",
            "format" : "int32"
          },
          "parameter" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "details" : {
            "uniqueItems" : true,
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ErrorDetails"
            }
          }
        }
      },
      "PurchaseRequestDto" : {
        "required" : [ "amount", "description", "transactionDate" ],
        "type" : "object",
        "properties" : {
          "description" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string",
            "description" : "Description"
          },
          "transactionDate" : {
            "type" : "string",
            "description" : "Transaction date",
            "format" : "date"
          },
          "amount" : {
            "type" : "number",
            "description" : "Amount"
          }
        }
      }
    }
  }
}